pod_containers/gateway/requirements.txt
uvicorn==0.35.0
requests==2.32.3
fastapi==0.115.14
httpx==0.28.1
python-dotenv==1.0.1
redisvl
snowflake-connector-python
